kmu-ems/backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.2.0
flask-restx==1.3.0
pymodbus==3.9.2
schedule==1.2.2
PyJWT==2.8.0
bcrypt==4.0.1
requests
datetime